# api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: WKF
host: localhost:10010
basePath: /api
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  "/files/{fileId}":
    x-swagger-router-controller: files
    delete:
      description: Delete file
      operationId: deleteFilesInBizsky
      parameters:
        - name: fileId
          description: fileId
          in: path
          type: string
          required: true
      responses:
        "200":
          description: Success
  "/import":
    x-swagger-router-controller: files
    post:
      operationId: postImportFile
      parameters:
        - in: "body"
          name: application
          description: The application to create.
          required: true
          schema:
            type: object
      responses:
        200:
          description: Logout
...
  '/files':
    x-swagger-router-controller: files
    post:
      summary: Uploads a file datalink.
      consumes:
        - multipart/form-data
      operationId: uploadDataLink
      parameters:
        - in: formData
          name: upfile
          type: file
          required: true
          description: The file to upload datalink.
      responses:
        '200':
          description: OK
    get:
      description: get list files from bizsky.
      operationId: getFiles
      responses:
        "200":
          description: OK
...